requirements.txt
asgiref==3.3.4
Django==3.2.4
django-adminlte3==0.1.6
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.12.4
django-filter~=2.4.0